+++ upcloud/service/retry_test.go
@@ -12,14 +12,14 @@ func TestRetry_noInverse(t *testing.T) {
 	t.Parallel()
 
 	ctx := context.TODO()
-	value, err := retry(ctx, func(i int, _ context.Context) (*string, error) {
+	value, err := Retry(ctx, func(i int, _ context.Context) (*string, error) {
 		if i < 3 {
 			return nil, nil
 		}
 
 		value := "ready"
 		return &value, nil
-	}, &retryConfig{interval: time.Millisecond * 125})
+	}, &RetryConfig{interval: time.Millisecond * 125})
 
 	assert.NoError(t, err)
 	assert.Equal(t, "ready", *value)
@@ -31,9 +31,9 @@ func TestRetry_timeout(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 500*time.Millisecond)
 	defer cancel()
 
-	value, err := retry(ctx, func(i int, _ context.Context) (*string, error) {
+	value, err := Retry(ctx, func(i int, _ context.Context) (*string, error) {
 		return nil, nil
-	}, &retryConfig{interval: time.Millisecond * 125})
+	}, &RetryConfig{interval: time.Millisecond * 125})
 
 	assert.Error(t, err)
 	assert.Nil(t, value)
@@ -43,12 +43,12 @@ func TestRetry_inverse(t *testing.T) {
 	t.Parallel()
 
 	ctx := context.TODO()
-	value, err := retry(ctx, func(i int, _ context.Context) (*int, error) {
+	value, err := Retry(ctx, func(i int, _ context.Context) (*int, error) {
 		if i < 3 {
 			return &i, nil
 		}
 		return nil, nil
-	}, &retryConfig{interval: time.Millisecond * 125, inverse: true})
+	}, &RetryConfig{interval: time.Millisecond * 125, inverse: true})
 
 	assert.NoError(t, err)
 	assert.Nil(t, value)
@@ -59,9 +59,9 @@ func TestRetry_noInterval(t *testing.T) {
 
 	tests := []struct {
 		name   string
-		config *retryConfig
+		config *RetryConfig
 	}{
-		{name: "no interval", config: &retryConfig{inverse: false}},
+		{name: "no interval", config: &RetryConfig{inverse: false}},
 		{name: "nil config", config: nil},
 	}
 
@@ -72,7 +72,7 @@ func TestRetry_noInterval(t *testing.T) {
 			ctx, cancel := context.WithTimeout(context.Background(), 125*time.Millisecond)
 			defer cancel()
 
-			value, err := retry(ctx, func(i int, _ context.Context) (*int, error) {
+			value, err := Retry(ctx, func(i int, _ context.Context) (*int, error) {
 				return &i, nil
 			}, test.config)